+++ R/scales.R
@@ -20,7 +20,7 @@
 #'   [ggplot2::continuous_scale()]
 #'
 scale_color_bde_d <- function(palette = "bde_vivid_pal",
-                              ...) {
+                              ...) { # nocov start
   valid_pals <- c("bde_vivid_pal", "bde_rose_pal")
 
   if (!palette %in% valid_pals) {
@@ -37,12 +37,13 @@ scale_color_bde_d <- function(palette = "bde_vivid_pal",
     palette = cols,
     ...
   )
+  # nocov end
 }
 
 #' @rdname scales_bde
 #' @export
 scale_fill_bde_d <- function(palette = "bde_vivid_pal",
-                             ...) {
+                             ...) { # nocov start
   valid_pals <- c("bde_vivid_pal", "bde_rose_pal")
 
   if (!palette %in% valid_pals) {
@@ -59,13 +60,14 @@ scale_fill_bde_d <- function(palette = "bde_vivid_pal",
     palette = cols,
     ...
   )
+  # nocov end
 }
 
 
 #' @rdname scales_bde
 #' @export
 scale_color_bde_c <- function(palette = "bde_vivid_pal",
-                              ...) {
+                              ...) { # nocov start
   valid_pals <- c("bde_vivid_pal", "bde_rose_pal")
 
   if (!palette %in% valid_pals) {
@@ -82,12 +84,13 @@ scale_color_bde_c <- function(palette = "bde_vivid_pal",
     palette = scales::gradient_n_pal(cols),
     ...
   )
+  # nocov end
 }
 
 #' @rdname scales_bde
 #' @export
 scale_fill_bde_c <- function(palette = "bde_vivid_pal",
-                             ...) {
+                             ...) { # nocov start
   valid_pals <- c("bde_vivid_pal", "bde_rose_pal")
 
   if (!palette %in% valid_pals) {
@@ -104,4 +107,5 @@ scale_fill_bde_c <- function(palette = "bde_vivid_pal",
     palette = scales::gradient_n_pal(cols),
     ...
   )
+  # nocov end
 }