+++ inst/tinytest/test_indicators.R
@@ -12,4 +12,6 @@ if (run_tests) {
   expect_silent(bde_ind_euribor_12m_daily())
   expect_silent(bde_ind_cpi_var())
   expect_silent(bde_ind_ibex())
+  expect_silent(bde_ind_gdp_quarterly())
+  expect_silent(bde_ind_population())
 }