+++ inst/tinytest/test_series.R
@@ -6,21 +6,24 @@ expect_null(bde_series_full_load("aa"))
 
 if (run_tests) {
   expect_message(bde_series_full_load("TI_1_1.csv",
-                                      cache_dir = tempdir(),
-                                      verbose = TRUE))
+    cache_dir = tempdir(),
+    verbose = TRUE
+  ))
   expect_silent(bde_series_full_load("TI_1_1.csv",
-                                     cache_dir = NULL,
-                                     verbose = TRUE))
+    cache_dir = NULL,
+    verbose = TRUE
+  ))
   expect_silent(bde_series_full_load("CF0101.csv",
-                                     cache_dir = NULL,
-                                     verbose = TRUE))
+    cache_dir = NULL,
+    verbose = TRUE
+  ))
   expect_silent(bde_series_full_load("CF0101"))
-  
-  
+
+
   data <- bde_series_full_load("TI_1_1.csv")
   meta <-
     bde_series_full_load("TI_1_1.csv", extract_metadata = TRUE)
-  
+
   expect_true(nrow(data) > nrow(meta))
 }
 
@@ -33,12 +36,12 @@ expect_error(bde_series_load(c(573234, 573234), series_label = c("a", "b")))
 
 if (run_tests) {
   expect_silent(bde_series_load(c(573234, 573214), series_label = c("a", "b")))
-  
+
   expect_silent(bde_series_load(573234, series_label = "a"))
   expect_silent(bde_series_load(573234, series_label = NULL))
   expect_silent(bde_series_load(573234, extract_metadata = TRUE))
   expect_message(bde_series_load(573234, verbose = TRUE))
-  
+
   meta <- bde_series_load(573234, extract_metadata = TRUE)
   data <- bde_series_load(573234, extract_metadata = FALSE)
   expect_true(nrow(data) > nrow(meta))