+++ NEWS.md
@@ -1,5 +1,8 @@
 # tidyBdE (development version)
 -   Add series index article.
+-   Add new series: `bde_ind_gdp_quarterly()`, `bde_ind_population()`.
+-   Improve package coverage and docs.
+
 
 # tidyBdE 0.1.2